requirements.txt
streamlit>=1.44.0
langchain>=1.2.0
langchain-google-genai>=4.2.0
langgraph>=1.1.0
google-generativeai>=0.8.0
chromadb>=1.5.0
faiss-cpu>=1.13.0
pymupdf>=1.24.0
pandas>=2.0.0
openpyxl>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
supabase>=2.0.0
streamlit-supabase-auth>=0.3.0
psycopg2-binary>=2.9.0
